requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.14
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
ollama==0.5.3
openai==1.99.9
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
